move theme controller
